requirements.txt
Django
djangorestframework
gunicorn
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
requests
torch
ultralytics
